Setup.py move and re-write. Added gmpy2 package.
